Get rid of TemporaryInterface and the Temporary trait, in favor of just using CompletionInterface.  Rename TemporaryManager to Temporary.
