+++ src/Contract/CompletionInterface.php
@@ -8,11 +8,12 @@ namespace Robo\Contract;
  *
  * Tasks that should be cleaned up when the program
  * terminates whenever they are used outside of a
- * task collection should be wrapped in TemporaryManager::temporaryTask().
- * @see Robo\Task\FileSystem\loadTasks::taskTmpDir
+ * task collection should be wrapped in
+ * Temporary::temporaryTask().  This will cause their
+ * complete() method to be called at shutdown time, but
+ * only if the object is not added to some other collection.
  *
- * Tasks that implement RollbackInterface should consider
- * implementing TemporaryInterface instead.
+ * @see Robo\Task\FileSystem\loadTasks::taskTmpDir
  *
  * Interface CompletionInterface
  * @package Robo\Contract
