+++ src/Contract/RollbackInterface.php
@@ -6,14 +6,6 @@ namespace Robo\Contract;
  * be called when the task collection it is added to
  * fails, and runs its rollback operation.
  *
- * Rollback operations can also be managed as temporary
- * tasks, just as in the CompletionInterface. However,
- * in this instance, the rollback operation will only
- * be done if TemporaryManager::fail() is explicitly called.
- *
- * Tasks that implement RollbackInterface should consider
- * implementing TemporaryInterface instead.
- *
  * Interface RollbackInterface
  * @package Robo\Contract
  */
