+++ src/Task/FileSystem/loadShortcuts.php
@@ -1,7 +1,7 @@
 <?php
 namespace Robo\Task\FileSystem;
 
-use Robo\Collection\TemporaryManager;
+use Robo\Collection\Temporary;
 
 trait loadShortcuts
 {
@@ -68,7 +68,7 @@ trait loadShortcuts
      */
     protected function _tmpDir($prefix = 'tmp', $base = '', $includeRandomPart = true)
     {
-        $result = TemporaryManager::temporaryTask(new TmpDir($prefix, $base, $includeRandomPart))->run();
+        $result = Temporary::temporaryTask(new TmpDir($prefix, $base, $includeRandomPart))->run();
         $data = $result->getData() + ['path' => ''];
         return $data['path'];
     }
